Get rid of unnecessary hierarcy structure
